Refactor: Moved matches screen into its own component
