+++ frontend/src/screens/Chats/Messages.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { useState } from 'react';
-import { View, Text, StyleSheet, SafeAreaView, ScrollView } from 'react-native';
+import { SafeAreaView, ScrollView } from 'react-native';
 import MessagesTabs from '../../components/ChatsComponents/MessagesTabs';
 import ChatList from '../../components/ChatsComponents/ChatList';
+import MatchList from '../../components/ChatsComponents/MatchList';
 
 const Messages = () => {
 
@@ -16,11 +17,7 @@ const Messages = () => {
         }
         else if(activeTab === 'matches') {
             return (
-                <View style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', marginTop: 280 }}>
-                    <Text style={styles.chatPage}>
-                        MATCHES SCREEN
-                    </Text>
-                </View>
+                <MatchList />
             );
         };
     };
@@ -34,13 +31,5 @@ const Messages = () => {
         </SafeAreaView>
     );
 };
-
-const styles = StyleSheet.create({
-    chatPage: {
-        fontFamily: 'Bebas Neue',
-        color: 'white',
-        fontSize: 50,
-    },
-});
 
 export default Messages;
